refactor: App 컴포넌트에서 상태와 카트 아이템 로직을 useShoppingCart 훅으로 분리
